Tune selecting the nearest other selection after remove
I could not remove empty selections, where `.a == .b`.  The previous
algorithm had no notion of having the cursor *on* a selection.

The algorithm now just sorts or ranks all saved selection by distance,
and then we choose the top one.

I don't know yet what `"mark"` is for so I did not change that at all
and calculate the "last_sel" just for that purpose.