Add one arbitrary exception to the logic
If you jump around or activate the saved selections, the plugin always
adds the current selection, the one on `view.sel()`.

This is usually great but make one exception:  if the current selection
is just a cursor but the saved selections are all not empty.

This is useful if adding the last cursor was a mistake and saves the
user from calling `power_cursors_remove` before calling the next
command.  (It just saves one keystroke.)